w5c: ueventd: Add /vendor/firmware_mnt/image/ to firmware_directories list
In an attempt to drop the legacy and unstable /firmware symlink once and
for all, enable ueventd to read firmware directly from our own firmware
mountpoint. This much safer than relying on a (SAR) image to specify the
link for us, which is not mandatory at all and requires nasty sepolicy
hacks.

Note that the path has to end with a slash because ueventd doesn't
intersperse a "/" inbetween while searching for firmware.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>